+++ Sources/Sentry/include/HybridPublic/SentryFramesTracker.h
@@ -20,6 +20,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  * Tracks total, frozen and slow frames for iOS, tvOS, and Mac Catalyst.
+ *
+ * @discussion This class ignores a couple of methods for the thread sanitizer. We intentionally
+ * accept several data races in this class, a decision that is driven by the fact that the code
+ * always writes on the main thread. This approach, while it may not provide 100% correct frame
+ * statistic for background spans, significantly reduces the overhead of synchronization, thereby
+ * enhancing performance.
  */
 @interface SentryFramesTracker : NSObject